{{cookiecutter.project_slug}}/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile '.\requirements-dev.in'
#
alabaster==0.7.13
    # via sphinx
asgiref==3.7.2
    # via
    #   -c .\requirements.txt
    #   django
autoflake==2.1.1
    # via -r .\requirements-dev.in
autopep8==2.0.2
    # via -r .\requirements-dev.in
babel==2.12.1
    # via sphinx
beautifulsoup4==4.12.2
    # via -r .\requirements-dev.in
black==23.3.0
    # via -r .\requirements-dev.in
certifi==2023.5.7
    # via
    #   -c .\requirements.txt
    #   requests
charset-normalizer==3.1.0
    # via
    #   -c .\requirements.txt
    #   requests
click==8.1.3
    # via
    #   black
    #   djlint
    #   mkdocs
colorama==0.4.6
    # via
    #   click
    #   djlint
    #   mkdocs
    #   mkdocs-material
    #   pytest
    #   sphinx
    #   taskipy
    #   tqdm
coverage[toml]==7.2.7
    # via pytest-cov
cssbeautifier==1.14.8
    # via djlint
django==4.2.2
    # via
    #   -c .\requirements.txt
    #   django-debug-toolbar
    #   django-extensions
    #   model-bakery
django-debug-toolbar==4.0.0
    # via -r .\requirements-dev.in
django-extensions==3.2.3
    # via -r .\requirements-dev.in
djlint==1.25.1
    # via -r .\requirements-dev.in
docutils==0.18.1
    # via
    #   sphinx
    #   sphinx-rtd-theme
editorconfig==0.12.3
    # via
    #   cssbeautifier
    #   jsbeautifier
exceptiongroup==1.1.1
    # via pytest
faker==18.6.2
    # via -r .\requirements-dev.in
ghp-import==2.1.0
    # via mkdocs
html-tag-names==0.1.2
    # via djlint
html-void-elements==0.1.0
    # via djlint
html5lib==1.1
    # via -r .\requirements-dev.in
idna==3.4
    # via
    #   -c .\requirements.txt
    #   requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r .\requirements-dev.in
jinja2==3.1.2
    # via
    #   -c .\requirements.txt
    #   mkdocs
    #   mkdocs-material
    #   mkdocstrings
    #   sphinx
jsbeautifier==1.14.8
    # via
    #   cssbeautifier
    #   djlint
json5==0.9.14
    # via djlint
markdown==3.3.7
    # via
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocs-material
    #   mkdocstrings
    #   pymdown-extensions
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.3
    # via
    #   -c .\requirements.txt
    #   jinja2
    #   mkdocstrings
mdurl==0.1.2
    # via markdown-it-py
mergedeep==1.3.4
    # via mkdocs
mkdocs==1.4.3
    # via
    #   -r .\requirements-dev.in
    #   mkdocs-autorefs
    #   mkdocs-material
    #   mkdocstrings
mkdocs-autorefs==0.4.1
    # via mkdocstrings
mkdocs-material==9.1.15
    # via -r .\requirements-dev.in
mkdocs-material-extensions==1.1.1
    # via mkdocs-material
mkdocstrings==0.21.2
    # via -r .\requirements-dev.in
model-bakery==1.11.0
    # via -r .\requirements-dev.in
mslex==0.3.0
    # via taskipy
mypy-extensions==1.0.0
    # via black
packaging==23.1
    # via
    #   -c .\requirements.txt
    #   black
    #   mkdocs
    #   pytest
    #   sphinx
pathspec==0.11.1
    # via
    #   black
    #   djlint
platformdirs==3.5.3
    # via black
pluggy==1.0.0
    # via pytest
psutil==5.9.5
    # via taskipy
pycodestyle==2.10.0
    # via autopep8
pyflakes==3.0.1
    # via autoflake
pygments==2.15.1
    # via
    #   mkdocs-material
    #   rich
    #   sphinx
pymdown-extensions==10.0.1
    # via
    #   mkdocs-material
    #   mkdocstrings
pytest==7.3.2
    # via
    #   -r .\requirements-dev.in
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r .\requirements-dev.in
pytest-django==4.5.2
    # via -r .\requirements-dev.in
python-dateutil==2.8.2
    # via
    #   -c .\requirements.txt
    #   faker
    #   ghp-import
pyyaml==6.0
    # via
    #   djlint
    #   mkdocs
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==0.1
    # via mkdocs
regex==2023.6.3
    # via
    #   -c .\requirements.txt
    #   djlint
    #   mkdocs-material
requests==2.31.0
    # via
    #   -c .\requirements.txt
    #   mkdocs-material
    #   sphinx
rich==13.3.5
    # via -r .\requirements-dev.in
six==1.16.0
    # via
    #   -c .\requirements.txt
    #   cssbeautifier
    #   html5lib
    #   jsbeautifier
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.4.1
    # via beautifulsoup4
sphinx==6.2.1
    # via
    #   -r requirements-dev.in
    #   sphinx-rtd-theme
    #   sphinxcontrib-jquery
sphinx-rtd-theme==1.2.2
    # via -r requirements-dev.in
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jquery==4.1
    # via sphinx-rtd-theme
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.4
    # via
    #   -c .\requirements.txt
    #   django
    #   django-debug-toolbar
taskipy==1.10.4
    # via -r .\requirements-dev.in
tomli==2.0.1
    # via
    #   autoflake
    #   autopep8
    #   black
    #   coverage
    #   djlint
    #   pytest
    #   taskipy
tqdm==4.65.0
    # via djlint
typing-extensions==4.6.3
    # via
    #   -c requirements.txt
    #   asgiref
tzdata==2023.3
    # via
    #   -c .\requirements.txt
    #   django
urllib3==2.0.3
    # via
    #   -c .\requirements.txt
    #   requests
validate-docbr==1.10.0
    # via -r .\requirements-dev.in
watchdog==3.0.0
    # via mkdocs
webencodings==0.5.1
    # via html5lib
